Refactor versioning classes to avoid circular dependencies
